Optimize Docker Command tests to one VCH (#6481)
Migrate the tests to using a single VCH if running as a group.
There are still some test files that install their a new VCH.
Some of those tests should be re-examined to determine if they
really need a new VCH.

A bootstrap and cleanup robot file is required to initiate and
clean up a group of tests. The infrastructure does not prevent
single test file from running normally.

To initiate "one VCH" mode, REUSE-VCH env variable must be
set to the VCH name in the bootstrap file. Also, the VCH name
must be added to the removal exception list.

Added DEBUG_VCH support to allow local test runs to use an
existing VCH created outside of the tests.  Simply create a
VCH using vic-machine and define the DEBUG_VCH in your secrets
file that names the VCH.

Enabling this mode reverts group 1 tests back to a VCH per suite
mode.  If this variable doesn't exist, the group t tests uses
a single shared VCH.  If TARGET_VCH is specified, MULTI_VCH is
ignored.  TARGET_VCH replaces the functionality of the previous
DEBUG_VCH.  Now, DEBUG_VCH toggles whether to dump out docker
images and running containers on a VCH at the start of a suite.

1. Added exception list to prevent VCH from getting "cleaned up"
2. Modify cleanup code to ignore assets linked to the exception list
3. Clean up containers with docker before cleaning up VIC appliance
4. Fix tests with dependencies on it running on its own VCH
5. Added a few docker utility keywords.
6. Add ability to use VCH created outside of CI
7. Added TARGET_VCH, MULTI_VCH, and DEBUG_VCH variables

Resolves #6301