Generate internal events on state update [full ci] (#7140)
* Generate internal events on state update

We were previously updating recorded container state during any refresh of
back end state; this was done to ensure we update internal state to
reflect backend state in the case of dropped incremental event updates.
However we were not triggering the internal events that should have fired
from the state transitions, which means certain management operations were
not occurring and docker API was not sending events to the client.

This makes state handling via events aware of the initialization status of
the containerVM and is also much more specific as to when state refreshes
are required.

NOTE:
If stopping and starting in very quick succession it's possible to receive
the VmStopped event from vSphere after we have started the VM again. This
could actually unbind a correctly configured network, but more likely will
result in a ConcurrentModificationError that will cause either the event
to be dropped or the start to be replayed. Either of those outcomes would be
correct but at the expense of more backend operations, however there is still
a window for the third incorrect path where no ConcurrentModificationError is
raised.


* Always run state update and always update start time

The status conversion in the personality is gated on start time being
non-nil. It's unclear to me why that is at this time so I have simply
ensured that starttime is set appropriately during out-of-band power
operations.

* Correct event handling and dispatch

This change causes out-of-band power on to transition through
Created->Starting->Running the same way an inline power on does.
In order to generate the correct series of docker events the "started"
event is now generated on receipt of the "Running" event from the
portlayer instead of as a presumed occurrance at the end of containerStart
in the personality. This has the added benefit that the event will not be
generated if the process fails to launch in the cVM, aligning with docker
behaviour.

The primary outstanding item of concern is that a ContainerList or
ContainerInspect operation is required to trigger the transition from
Starting->Running.
This is a compromise because the complexity of adding a proper delayed
trigger and integrating it with existing WaitFor... paths is way to high
for a point release.

The tests have been updated for the out-of-band power on case to include
a specific IP address - if this isn't done the the container has no address
bound and will attempt to acquire a DHCP address on the bridge; an
exercise doomed to failure in a correctly configured environment.